+++ src/anari_extension_utility.cpp
@@ -1,0 +1,12 @@
+#define ANARI_EXTENSION_UTILITY_IMPL
+
+// TODO: Fix upstream
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include "anari/frontend/anari_extension_utility.h"
+
+#ifdef __cplusplus
+}
+#endif