+++ src/type_utility.cpp
@@ -1,0 +1,4 @@
+// TODO: Add ANARI_INLINE upstream
+#define inline
+
+#include "anari/frontend/type_utility.h"